Move filters directory inside geoengine where it belongs
